COMP and CLR. (#24)
* Refactor numeric type mapping by introducing IMapper interface and specific mappers for signed and unsigned integers

* Refactor numeric encoding and decoding by implementing a new mapping system with NumericMapper, enhancing exception handling for digit limits and removing deprecated mappers.

* Updated namespaces for numeric encoders and decoders to use a more structured approach.

+++ GetThePicture.Tests/Picture/Clause/Comp3Test.cs
@@ -43,10 +43,10 @@ public class Comp3Test
     // ------------------------- 
 
     [DataTestMethod]
-    [DataRow( "9(5)",  52194UL, new byte[] { 0x52, 0x19, 0x4F })]
-    [DataRow( "9(5)",  52194UL, new byte[] { 0x52, 0x19, 0x4C })] // TODO: Sign Nibble 與 PIC 定義不合，看看要不要擋
-    [DataRow("S9(5)",  52194L , new byte[] { 0x52, 0x19, 0x4C })]
-    [DataRow("S9(5)", -52194L , new byte[] { 0x52, 0x19, 0x4D })]
+    [DataRow( "9(5)",(uint)  52194, new byte[] { 0x52, 0x19, 0x4F })]
+    [DataRow( "9(5)",(uint)  52194, new byte[] { 0x52, 0x19, 0x4C })] // TODO: Sign Nibble 與 PIC 定義不合，看看要不要擋
+    [DataRow("S9(5)",(int)   52194, new byte[] { 0x52, 0x19, 0x4C })]
+    [DataRow("S9(5)",(int)  -52194, new byte[] { 0x52, 0x19, 0x4D })]
     public void Decode_Combination_Test(string picString, object expected, byte[] buffer)
     {
         var pic = PicMeta.Parse(picString);
@@ -67,7 +67,7 @@ public class Comp3Test
 
         var value = PicClauseCodec.ForMeta(pic).Decode([0x52, 0x19, 0x4D]);
 
-        Assert.AreEqual(-194L, value);
+        Assert.AreEqual((short) -194, value);
     }
 
     // -------------------------
